Add the rust-wasm-book
This book seeks to aggregate various documents into one place making it
easy to display the documentation online much like we do with TRPL and
to make a central area for writing those documents.

This splits up the `workflows.md` file into a basic hello-world chapter,
a modified setup chapter, and then an actual workflows chapter that can
be worked on with things that are more like workflows than examples per
se.

The README has been update on how to build the book and what the book is
about for those wishing to contribute to those docs.