Call drawGrid() and drawCells() prior to the first invocation of tick().
This commit adds a `drawGrid()` and `drawCells()` call just before the
first call to `requestAnimationFrame()` in order to ensure that the initial
state of the universe is drawn prior to executing a `Universe.tick()` call.

If this is not done, the initial state of the universe is never drawn,
which can be confusing for those looking to implement the exercises where a
specific starting set of cells is desired.

Fixes #118.